Relocate shader builder
