+++ game/shader_exposure.go
@@ -1,13 +1,13 @@
 package game
 
 import (
-	"github.com/mokiat/lacking-js/webgl"
+	"github.com/mokiat/lacking-js/internal"
 	"github.com/mokiat/lacking/game/graphics"
 )
 
 func newExposureShaderSet() graphics.ShaderSet {
-	vsBuilder := webgl.NewShaderSourceBuilder(exposureVertexShader)
-	fsBuilder := webgl.NewShaderSourceBuilder(exposureFragmentShader)
+	vsBuilder := internal.NewShaderSourceBuilder(exposureVertexShader)
+	fsBuilder := internal.NewShaderSourceBuilder(exposureFragmentShader)
 	return graphics.ShaderSet{
 		VertexShader:   vsBuilder.Build,
 		FragmentShader: fsBuilder.Build,
