+++ game/shader_skybox.go
@@ -1,13 +1,13 @@
 package game
 
 import (
-	"github.com/mokiat/lacking-js/webgl"
+	"github.com/mokiat/lacking-js/internal"
 	"github.com/mokiat/lacking/game/graphics"
 )
 
 func newSkyboxShaderSet() graphics.ShaderSet {
-	vsBuilder := webgl.NewShaderSourceBuilder(cubeSkyboxVertexShader)
-	fsBuilder := webgl.NewShaderSourceBuilder(cubeSkyboxFragmentShader)
+	vsBuilder := internal.NewShaderSourceBuilder(cubeSkyboxVertexShader)
+	fsBuilder := internal.NewShaderSourceBuilder(cubeSkyboxFragmentShader)
 	return graphics.ShaderSet{
 		VertexShader:   vsBuilder.Build,
 		FragmentShader: fsBuilder.Build,
