+++ game/shader_skycolor.go
@@ -1,13 +1,13 @@
 package game
 
 import (
-	"github.com/mokiat/lacking-js/webgl"
+	"github.com/mokiat/lacking-js/internal"
 	"github.com/mokiat/lacking/game/graphics"
 )
 
 func newSkycolorShaderSet() graphics.ShaderSet {
-	vsBuilder := webgl.NewShaderSourceBuilder(colorSkyboxVertexShader)
-	fsBuilder := webgl.NewShaderSourceBuilder(colorSkyboxFragmentShader)
+	vsBuilder := internal.NewShaderSourceBuilder(colorSkyboxVertexShader)
+	fsBuilder := internal.NewShaderSourceBuilder(colorSkyboxFragmentShader)
 	return graphics.ShaderSet{
 		VertexShader:   vsBuilder.Build,
 		FragmentShader: fsBuilder.Build,
