+++ ui/shader_text.go
@@ -1,18 +1,18 @@
 package ui
 
 import (
-	"github.com/mokiat/lacking-js/webgl"
+	"github.com/mokiat/lacking-js/internal"
 	"github.com/mokiat/lacking/ui"
 )
 
 func newTextShaders() ui.ShaderSet {
 	return ui.ShaderSet{
 		VertexShader: func() string {
-			builder := webgl.NewShaderSourceBuilder(textMaterialVertexShaderTemplate)
+			builder := internal.NewShaderSourceBuilder(textMaterialVertexShaderTemplate)
 			return builder.Build()
 		},
 		FragmentShader: func() string {
-			builder := webgl.NewShaderSourceBuilder(textMaterialFragmentShaderTemplate)
+			builder := internal.NewShaderSourceBuilder(textMaterialFragmentShaderTemplate)
 			return builder.Build()
 		},
 	}
